requirements.txt
pandas
requests
psycopg2-binary
pytz
